+++ src/containers/Activation.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import { ActivePage } from 'twreporter-registration'
+import { browserHistory } from 'react-router'
+
+const Activation = (props) => (
+  <div>
+    <ActivePage
+      activateRedirectPath={'/'}
+      browserHistory={browserHistory}
+      {...props}
+    />
+  </div>
+)
+
+export default Activation