+++ src/containers/App.js
@@ -5,6 +5,7 @@ import zhLocaleData from 'react-intl/locale-data/zh'
 // import locale data
 import { addLocaleData, IntlProvider } from 'react-intl'
 import { connect } from 'react-redux'
+import { signOutAction } from 'twreporter-registration'
 import Layout from '../components/Layout'
 
 // lodash
@@ -36,12 +37,13 @@ class App extends PureComponent {
 
   render() {
     const pathname = _.get(this.props, 'location.pathname')
-
     return (
       <IntlProvider locale={currentLocale} defaultLocale="zh-Hant">
         <Layout
           header={this.props.header}
           pathname={pathname}
+          signOutAction={this.props.signOutAction}
+          ifAuthenticated={this.props.ifAuthenticated}
         >
           {this.props.children}
         </Layout>
@@ -54,10 +56,21 @@ App.childContextTypes = {
   location: React.PropTypes.object
 }
 
+App.defaultProps = {
+  signOutAction: () => {},
+  ifAuthenticated: false
+}
+
+App.propTypes = {
+  signOutAction: React.PropTypes.func,
+  ifAuthenticated: React.PropTypes.bool
+}
+
 function mapStateToProps(state) {
   return {
-    header: state.header
+    header: state.header,
+    ifAuthenticated: _.get(state, [ 'auth', 'authenticated' ], false)
   }
 }
 
-export default connect(mapStateToProps)(App)
+export default connect(mapStateToProps, { signOutAction })(App)