+++ src/utils/style-utils.js
@@ -1,0 +1,53 @@
+import { css } from 'styled-components'
+
+export const breakPoints = {
+  mobileMaxWidth: '767px',
+  tabletMinWidth: '768px',
+  tabletMaxWidth: '1023px',
+  desktopMinWidth: '1024px',
+  desktopMaxWidth: '1439px',
+  overDesktopMinWidth: '1440px'
+}
+
+export const mediaScreen = {
+  mobile: (...args) => css`
+    @media (max-width: ${breakPoints.mobileMaxWidth}) {
+      ${css(...args)}
+    }
+  `,
+  tabletAbove: (...args) => css`
+    @media (min-width: ${breakPoints.tabletMinWidth}) {
+      ${css(...args)}
+    }
+  `,
+  tablet: (...args) => css`
+    @media (min-width: ${breakPoints.tabletMinWidth}) and (max-width: ${breakPoints.tabletMaxWidth}) {
+      ${css(...args)}
+    }
+  `,
+  tabletBelow: (...args) => css`
+    @media (max-width: ${breakPoints.tabletMaxWidth}) {
+      ${css(...args)}
+    }
+  `,
+  desktopAbove: (...args) => css`
+    @media (min-width: ${breakPoints.desktopMinWidth}) {
+      ${css(...args)}
+    }
+  `,
+  desktop: (...args) => css`
+    @media (min-width: ${breakPoints.desktopMinWidth}) and (max-width: ${breakPoints.desktopMaxWidth}) {
+      ${css(...args)}
+    }
+  `,
+  desktopBelow: (...args) => css`
+    @media (max-width: ${breakPoints.desktopMaxWidth}) {
+      ${css(...args)}
+    }
+  `,
+  overDesktop: (...args) => css`
+    @media (min-width: ${breakPoints.overDesktopMinWidth}) {
+      ${css(...args)}
+    }
+  `
+}